Extract proposal builder domain panels
