+++ src/features/proposals/components/proposal-simulate-form.tsx
@@ -6,7 +6,7 @@ import { Controller, useForm } from "react-hook-form";
 import { useQuery } from "@tanstack/react-query";
 import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Alert, Button, MenuItem, Stack, TextField } from "@mui/material";
+import { Alert, Button, Stack, TextField } from "@mui/material";
 
 import { getPortfolioBook, getPortfolioWorkspaceShell } from "@/apps/portfolio/api";
 import type { PortfolioPositionView } from "@/apps/portfolio/types";
@@ -23,13 +23,19 @@ import {
   createTradeIntent,
   createTradeIntentFromPosition,
   formatCurrencyValue,
-  formatPercentValue,
-  formatUnitValue,
   type ProposalDraftCashFlowIntent,
   type ProposalDraftTradeIntent,
 } from "../proposal-draft-preview";
 import type { AdvisoryWorkspaceEnvelopeResponse, ProposalSimulateResponse } from "../types";
-import { SectionBlock, Text } from "@/design-system";
+import { SectionBlock } from "@/design-system";
+import {
+  AdviseEvaluationSummaryPanel,
+  CashMovementsPanel,
+  CurrentPositionsPanel,
+  DraftOrderBlotterPanel,
+  IndicativeDraftImpactPanel,
+  SavedAdvisoryDraftPanel,
+} from "./proposal-builder-domain-panels";
 import styles from "./proposal-simulate-form.module.css";
 
 const schema = z.object({
@@ -515,295 +521,35 @@ export default function ProposalSimulateForm({
               </div>
             </section>
 
-            <section className={styles.panel} aria-labelledby="current-positions-heading">
-              <div className={styles.panelHeader}>
-                <div>
-                  <h3 id="current-positions-heading">Current Positions</h3>
-                  <p>
-                    Start from the live portfolio book, then buy more units or sell down holdings into
-                    the advisor-use draft.
-                  </p>
-                </div>
-                <span>{positionsLoading ? "Loading" : `${positions.length} positions`}</span>
-              </div>
-              {positions.length ? (
-                <div className={styles.positionsTableWrap}>
-                  <table className={styles.positionsTable}>
-                    <thead>
-                      <tr>
-                        <th>Instrument</th>
-                        <th>Asset Class</th>
-                        <th>Units</th>
-                        <th>Market Value</th>
-                        <th>Weight</th>
-                        <th>Draft Action</th>
-                      </tr>
-                    </thead>
-                    <tbody>
-                      {positions.slice(0, 12).map((position) => (
-                        <tr key={position.security_id}>
-                          <td>
-                            <strong>{position.instrument_name}</strong>
-                            <span>{position.security_id}</span>
-                          </td>
-                          <td>{position.asset_class ?? "Unclassified"}</td>
-                          <td>{formatUnitValue(position.quantity)}</td>
-                          <td>
-                            {formatCurrencyValue(position.market_value_base ?? 0, baseCurrency || "USD")}
-                          </td>
-                          <td>{formatPercentValue(position.weight_pct ?? 0)}</td>
-                          <td>
-                            <div className={styles.positionActions}>
-                              <Button
-                                type="button"
-                                size="small"
-                                variant="outlined"
-                                onClick={() => addPositionTrade(position, "BUY")}
-                              >
-                                Buy More
-                              </Button>
-                              <Button
-                                type="button"
-                                size="small"
-                                variant="outlined"
-                                color="inherit"
-                                onClick={() => addPositionTrade(position, "SELL")}
-                              >
-                                Sell Down
-                              </Button>
-                            </div>
-                          </td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                  {positions.length > 12 ? (
-                    <Text variant="metadata">
-                      Showing first 12 holdings. Use the instrument field below to add another held or off-book security.
-                    </Text>
-                  ) : null}
-                </div>
-              ) : (
-                <div className={styles.emptyBookNotice}>
-                  {positionsLoading
-                    ? "Loading current holdings from the portfolio book."
-                    : "No current positions are available. Add cash or an off-book instrument to begin the draft."}
-                </div>
-              )}
-            </section>
+            <CurrentPositionsPanel
+              positions={positions}
+              positionsLoading={positionsLoading}
+              baseCurrency={baseCurrency || "USD"}
+              onAddPositionTrade={addPositionTrade}
+            />
 
-            <section className={styles.panel} aria-labelledby="cash-movements-heading">
-              <div className={styles.panelHeader}>
-                <div>
-                  <h3 id="cash-movements-heading">Cash Movements</h3>
-                  <p>Model client subscriptions, withdrawals, and liquidity changes.</p>
-                </div>
-                <span>Net {netCashImpact()}</span>
-              </div>
-              <div className={styles.intentRows}>
-                {cashFlows.map((item) => (
-                  <div key={item.id} className={styles.cashRow}>
-                    <TextField
-                      label="Movement"
-                      size="small"
-                      select
-                      value={item.direction}
-                      onChange={(event) =>
-                        updateCashFlow(item.id, { direction: event.target.value as "IN" | "OUT" })
-                      }
-                    >
-                      <MenuItem value="IN">Inflow</MenuItem>
-                      <MenuItem value="OUT">Outflow</MenuItem>
-                    </TextField>
-                    <TextField
-                      label="Currency"
-                      size="small"
-                      value={item.currency}
-                      onChange={(event) => updateCashFlow(item.id, { currency: event.target.value })}
-                    />
-                    <TextField
-                      label="Amount"
-                      size="small"
-                      type="number"
-                      value={item.amount}
-                      onChange={(event) => {
-                        const next = (event.target as HTMLInputElement).valueAsNumber;
-                        updateCashFlow(item.id, { amount: Number.isNaN(next) ? 0 : next });
-                      }}
-                    />
-                    <TextField
-                      label="Advisor Note"
-                      size="small"
-                      fullWidth
-                      value={item.description ?? ""}
-                      onChange={(event) => updateCashFlow(item.id, { description: event.target.value })}
-                    />
-                    <Button
-                      type="button"
-                      variant="outlined"
-                      color="inherit"
-                      disabled={cashFlows.length === 1}
-                      onClick={() => setCashFlows((current) => current.filter((row) => row.id !== item.id))}
-                    >
-                      Remove
-                    </Button>
-                  </div>
-                ))}
-              </div>
-              <Button
-                type="button"
-                variant="outlined"
-                onClick={() => addCashFlow(form.getValues().baseCurrency || "USD")}
-              >
-                Add Cash Movement
-              </Button>
-            </section>
+            <CashMovementsPanel
+              cashFlows={cashFlows}
+              netCashImpact={netCashImpact()}
+              onUpdateCashFlow={updateCashFlow}
+              onRemoveCashFlow={(id) =>
+                setCashFlows((current) => current.filter((row) => row.id !== id))
+              }
+              onAddCashFlow={() => addCashFlow(form.getValues().baseCurrency || "USD")}
+            />
 
-            <section className={styles.panel} aria-labelledby="security-orders-heading">
-              <div className={styles.panelHeader}>
-                <div>
-                  <h3 id="security-orders-heading">Draft Order Blotter</h3>
-                  <p>
-                    Edit units for held positions or add an off-book instrument that is not already in
-                    the portfolio.
-                  </p>
-                </div>
-                <span>{validTradeCount()} ready</span>
-              </div>
-              <div className={styles.intentRows}>
-                {trades.map((item) => (
-                  <div key={item.id} className={styles.tradeRow}>
-                    <TextField
-                      label="Order"
-                      size="small"
-                      select
-                      value={item.side}
-                      onChange={(event) => updateTrade(item.id, { side: event.target.value as "BUY" | "SELL" })}
-                    >
-                      <MenuItem value="BUY">Buy</MenuItem>
-                      <MenuItem value="SELL">Sell</MenuItem>
-                    </TextField>
-                    <TextField
-                      label="Instrument"
-                      size="small"
-                      fullWidth
-                      value={item.instrumentId}
-                      onChange={(event) => updateTrade(item.id, { instrumentId: event.target.value })}
-                    />
-                    <TextField
-                      label="Asset Class"
-                      size="small"
-                      value={item.assetClass ?? ""}
-                      onChange={(event) => updateTrade(item.id, { assetClass: event.target.value })}
-                      placeholder="Equities"
-                    />
-                    <TextField
-                      label="Quantity"
-                      size="small"
-                      type="number"
-                      value={item.quantity}
-                      onChange={(event) => {
-                        const next = (event.target as HTMLInputElement).valueAsNumber;
-                        updateTrade(item.id, { quantity: Number.isNaN(next) ? 0 : next });
-                      }}
-                    />
-                    <TextField
-                      label="Reference Price"
-                      size="small"
-                      type="number"
-                      value={item.referencePrice ?? 0}
-                      onChange={(event) => {
-                        const next = (event.target as HTMLInputElement).valueAsNumber;
-                        updateTrade(item.id, { referencePrice: Number.isNaN(next) ? 0 : next });
-                      }}
-                      helperText={item.source === "NEW_INSTRUMENT" ? "Used for indicative preview" : undefined}
-                    />
-                    <Button
-                      type="button"
-                      variant="outlined"
-                      color="inherit"
-                      disabled={trades.length === 1}
-                      onClick={() => setTrades((current) => current.filter((row) => row.id !== item.id))}
-                    >
-                      Remove
-                    </Button>
-                  </div>
-                ))}
-              </div>
-              <Button type="button" variant="outlined" onClick={addTrade}>
-                Add Off-Book Instrument
-              </Button>
-            </section>
+            <DraftOrderBlotterPanel
+              trades={trades}
+              readyTradeCount={validTradeCount()}
+              onUpdateTrade={updateTrade}
+              onRemoveTrade={(id) => setTrades((current) => current.filter((row) => row.id !== id))}
+              onAddTrade={addTrade}
+            />
 
-            <section className={styles.panel} aria-labelledby="draft-impact-heading">
-              <div className={styles.panelHeader}>
-                <div>
-                  <h3 id="draft-impact-heading">Indicative Draft Impact</h3>
-                  <p>
-                    Live advisor preview from current holdings, draft orders, and cash movements.
-                    Formal suitability, risk, and allocation proof is produced by simulation.
-                  </p>
-                </div>
-                <span>
-                  {formatCurrencyValue(draftPreview.proposedPortfolioValue, baseCurrency || "USD")}
-                </span>
-              </div>
-              <div className={styles.impactSummaryGrid}>
-                <div>
-                  <span>Current Value</span>
-                  <strong>
-                    {formatCurrencyValue(draftPreview.currentPortfolioValue, baseCurrency || "USD")}
-                  </strong>
-                </div>
-                <div>
-                  <span>Proposed Value</span>
-                  <strong>
-                    {formatCurrencyValue(draftPreview.proposedPortfolioValue, baseCurrency || "USD")}
-                  </strong>
-                </div>
-                <div>
-                  <span>Largest Position</span>
-                  <strong>
-                    {formatPercentValue(draftPreview.currentLargestWeight)} →{" "}
-                    {formatPercentValue(draftPreview.proposedLargestWeight)}
-                  </strong>
-                </div>
-                <div>
-                  <span>Unpriced Draft Lines</span>
-                  <strong>{draftPreview.unpricedTradeCount}</strong>
-                </div>
-              </div>
-              <div className={styles.positionsTableWrap}>
-                <table className={styles.positionsTable}>
-                  <thead>
-                    <tr>
-                      <th>Asset Class</th>
-                      <th>Current Weight</th>
-                      <th>Proposed Weight</th>
-                      <th>Change</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {draftPreview.allocationRows.map((row) => (
-                      <tr key={row.assetClass}>
-                        <td>{row.assetClass}</td>
-                        <td>{formatPercentValue(row.currentWeight)}</td>
-                        <td>{formatPercentValue(row.proposedWeight)}</td>
-                        <td className={row.proposedWeight >= row.currentWeight ? styles.positive : styles.negative}>
-                          {row.proposedWeight >= row.currentWeight ? "+" : ""}
-                          {formatPercentValue(row.proposedWeight - row.currentWeight)}
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-              {draftPreview.unpricedTradeCount ? (
-                <Alert severity="warning">
-                  Add reference prices for all draft lines to complete the indicative allocation preview.
-                </Alert>
-              ) : null}
-            </section>
+            <IndicativeDraftImpactPanel
+              draftPreview={draftPreview}
+              baseCurrency={baseCurrency || "USD"}
+            />
 
             <section className={styles.panel} aria-labelledby="draft-details-heading">
               <div className={styles.panelHeader}>
@@ -839,65 +585,18 @@ export default function ProposalSimulateForm({
       ) : null}
 
       {result ? (
-        <section className={styles.resultPanel} aria-label="Simulation summary">
-          <Text variant="sectionTitle">Advise Evaluation Summary</Text>
-          <div className={styles.resultGrid}>
-            <div>
-              <span>Status</span>
-              <strong>{result.data.status ?? "UNKNOWN"}</strong>
-            </div>
-            <div>
-              <span>Proposal Run</span>
-              <strong>{String(result.data.proposal_run_id ?? "N/A")}</strong>
-            </div>
-            <div>
-              <span>Correlation</span>
-              <strong>{result.correlation_id}</strong>
-            </div>
-          </div>
-          {simulationHighlights(result).length ? (
-            <div className={styles.outputGrid}>
-              {simulationHighlights(result).map((highlight) => (
-                <div key={highlight.label}>
-                  <span>{highlight.label}</span>
-                  <strong>{highlight.value}</strong>
-                </div>
-              ))}
-            </div>
-          ) : (
-            <Text variant="secondary">No additional scalar metrics were returned by the simulation engine.</Text>
-          )}
-          {extractEvaluationSummary(workspaceEnvelope) ? (
-            <div className={styles.outputGrid}>
-              <div>
-                <span>Review Issues</span>
-                <strong>{String(extractEvaluationSummary(workspaceEnvelope)?.review_issue_count ?? 0)}</strong>
-              </div>
-              <div>
-                <span>Blocking Issues</span>
-                <strong>{String(extractEvaluationSummary(workspaceEnvelope)?.blocking_issue_count ?? 0)}</strong>
-              </div>
-              <div>
-                <span>Draft Trades</span>
-                <strong>
-                  {String(
-                    recordValue(extractEvaluationSummary(workspaceEnvelope)?.impact_summary)?.trade_count ?? 0
-                  )}
-                </strong>
-              </div>
-            </div>
-          ) : null}
-        </section>
+        <AdviseEvaluationSummaryPanel
+          result={result}
+          highlights={simulationHighlights(result)}
+          reviewIssueCount={extractEvaluationSummary(workspaceEnvelope)?.review_issue_count}
+          blockingIssueCount={extractEvaluationSummary(workspaceEnvelope)?.blocking_issue_count}
+          draftTradeCount={
+            recordValue(extractEvaluationSummary(workspaceEnvelope)?.impact_summary)?.trade_count
+          }
+        />
       ) : null}
 
-      {savedProposalId ? (
-        <section className={styles.resultPanel} aria-label="Saved advisory draft">
-          <Text variant="body">Draft saved as Proposal ID: {savedProposalId}</Text>
-          <Text variant="body">
-            <Link href={`/proposals/${savedProposalId}`}>Open Proposal Details</Link>
-          </Text>
-        </section>
-      ) : null}
+      {savedProposalId ? <SavedAdvisoryDraftPanel proposalId={savedProposalId} /> : null}
     </SectionBlock>
   );
 }
